refactor: provide better strategies exports
